Refactoring a lot
- feature: polymorphism for DB transaction using abstract class
- feature: hook up new refactored methods and classes
- feature: remove redundant code
