+++ TorchDonation/src/main/java/net/gooday2die/torchdonation/CommandHandler/Redeem.java
@@ -14,7 +14,6 @@ import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
-import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.scheduler.BukkitRunnable;
 
 import java.util.regex.Pattern;
