+++ TorchDonation/src/main/java/net/gooday2die/torchdonation/CulturelandDonation/RewardUser.java
@@ -8,16 +8,6 @@ package net.gooday2die.torchdonation.CulturelandDonation;
  * @author Gooday2die @ https://github.com/gooday2die/TorchDonation
  */
 
-import net.gooday2die.torchdonation.ConfigValues;
-import net.gooday2die.torchdonation.dbHandler.dbConnection;
-import org.bukkit.Bukkit;
-import org.bukkit.ChatColor;
-import org.bukkit.command.CommandSender;
-import org.bukkit.plugin.java.JavaPlugin;
-import org.checkerframework.checker.units.qual.C;
-
-import java.util.concurrent.ExecutionException;
-
 /**
  * A method for rewarding user when donation was successful.
  */
