--- TorchDonation/src/main/java/net/gooday2die/torchdonation/dbHandler/DBInterface.java
@@ -1,4 +1,0 @@
-package net.gooday2die.torchdonation.dbHandler;
-
-public interface DBInterface {
-}
